Improved UI Test Structure
